[style] restyle student eContact page and conditional rendering
