+++ app/navigation/StudentDashboardNavigation.js
@@ -1,7 +1,7 @@
 import { StackNavigator } from 'react-navigation';
 import StudentDashboard from '../components/student/StudentDashboard';
 import CheckIn from '../components/student/CheckIn';
-import EmergencyContact from '../components/teacher/EmergencyContact';
+import EmergencyContact from '../components/student/EmergencyContact';
 
 const StudentDashboardNavigation = StackNavigator({
   Home: {
