+++ app/navigation/TeacherDrawerNavigation.js
@@ -3,7 +3,7 @@ import TeacherDashboardNavigation from './TeacherDashboardNavigation';
 import TeacherClassNavigation from './TeacherClassNavigation';
 import TeacherDrawer from '../components/teacher/TeacherDrawer';
 import AddClass from '../components/teacher/AddClass';
-import EmergencyContact from '../components/teacher/EmergencyContact';
+import EmergencyContact from '../components/student/EmergencyContact';
 
 const TeacherDrawerNavigator = DrawerNavigator({
   TeacherDashboardNavigation: {
